# src/main/java/com/fidd/api/controller/TaskApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.fidd.api.controller;

import com.fidd.api.model.TaskDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.constraints.NotNull;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.List;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "Task", description = "the Task API")
public interface TaskApi
{

  default Optional<NativeWebRequest> getRequest()
  {
    return Optional.empty();
  }

  public static final String PATH_GET_USER_TASKS = "/users/{userId}/tasks";

  /**
   * GET /users/{userId}/tasks : Get tasks for a specific user
   *
   * @param userId (required)
   * @return List of user tasks (status code 200)
   */
  @Operation(
    operationId = "getUserTasks",
    summary = "Get tasks for a specific user",
    tags = {"Task"},
    responses = {
      @ApiResponse(responseCode = "200", description = "List of user tasks", content = {
        @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TaskDto.class)))
      })
    }
  )
  @RequestMapping(
    method = RequestMethod.GET,
    value = TaskApi.PATH_GET_USER_TASKS,
    produces = {"application/json"}
  )

  default ResponseEntity<List<TaskDto>> getUserTasks(
    @NotNull @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") Integer userId
  )
  {
    getRequest().ifPresent(request ->
    {
      for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept")))
      {
        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json")))
        {
          String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"Implement backend endpoints for user management\", \"id\" : 0, \"title\" : \"Implement API\", \"status\" : \"TODO\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"Implement backend endpoints for user management\", \"id\" : 0, \"title\" : \"Implement API\", \"status\" : \"TODO\" } ]";
          ApiUtil.setExampleResponse(request, "application/json", exampleString);
          break;
        }
      }
    });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

  }

}