# src/main/java/com/fidd/api/controller/UserApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.fidd.api.controller;

import com.fidd.api.model.UserCreateRequest;
import com.fidd.api.model.UserDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.List;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "User", description = "the User API")
public interface UserApi
{

  default Optional<NativeWebRequest> getRequest()
  {
    return Optional.empty();
  }

  public static final String PATH_CREATE_USER = "/users";

  /**
   * POST /users : Create new user
   *
   * @param userCreateRequest (required)
   * @return User created (status code 201)
   */
  @Operation(
    operationId = "createUser",
    summary = "Create new user",
    tags = {"User"},
    responses = {
      @ApiResponse(responseCode = "201", description = "User created", content = {
        @Content(mediaType = "application/json", schema = @Schema(implementation = UserDto.class))
      })
    }
  )
  @RequestMapping(
    method = RequestMethod.POST,
    value = UserApi.PATH_CREATE_USER,
    produces = {"application/json"},
    consumes = {"application/json"}
  )

  default ResponseEntity<UserDto> createUser(
    @Parameter(name = "UserCreateRequest", description = "", required = true) @Valid @RequestBody UserCreateRequest userCreateRequest
  )
  {
    getRequest().ifPresent(request ->
    {
      for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept")))
      {
        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json")))
        {
          String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"Ivan Petrov\", \"id\" : 1, \"email\" : \"ivan@example.com\" }";
          ApiUtil.setExampleResponse(request, "application/json", exampleString);
          break;
        }
      }
    });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

  }


  public static final String PATH_GET_USERS = "/users";

  /**
   * GET /users : Get list of users
   *
   * @return List of users (status code 200)
   */
  @Operation(
    operationId = "getUsers",
    summary = "Get list of users",
    tags = {"User"},
    responses = {
      @ApiResponse(responseCode = "200", description = "List of users", content = {
        @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserDto.class)))
      })
    }
  )
  @RequestMapping(
    method = RequestMethod.GET,
    value = UserApi.PATH_GET_USERS,
    produces = {"application/json"}
  )

  default ResponseEntity<List<UserDto>> getUsers(

  )
  {
    getRequest().ifPresent(request ->
    {
      for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept")))
      {
        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json")))
        {
          String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"Ivan Petrov\", \"id\" : 1, \"email\" : \"ivan@example.com\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"Ivan Petrov\", \"id\" : 1, \"email\" : \"ivan@example.com\" } ]";
          ApiUtil.setExampleResponse(request, "application/json", exampleString);
          break;
        }
      }
    });
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

  }

}